Migrate account dropdown menu (#262)
