+++ packages/frontend/src/locales/en.json
@@ -1,6 +1,5 @@
 {
   "brandText": "plumber",
-  "accountDropdownMenu.logout": "Logout",
   "drawer.dashboard": "Dashboard",
   "drawer.explore": "Explore",
   "app.addConnection": "Add connection",
